+++ packages/main/src/TextArea.ts
@@ -52,9 +52,9 @@ type IndexedTokenizedText = Array<{
 }>;
 
 type ExceededText = {
-	exceededText: string | undefined;
-	leftCharactersCount: number | undefined;
-	calcedMaxLength: number | undefined;
+	exceededText?: string;
+	leftCharactersCount?: number;
+	calcedMaxLength?: number;
 };
 
 /**
@@ -414,6 +414,11 @@ class TextArea extends UI5Element implements IFormElement {
 	}
 
 	onBeforeRendering() {
+		if (!this.value) {
+			// fallback to default value
+			this.value = "";
+		}
+
 		this._exceededTextProps = this._calcExceededText();
 		this._mirrorText = this._tokenizeText(this.value);
 
@@ -469,7 +474,7 @@ class TextArea extends UI5Element implements IFormElement {
 
 	_onfocusout(e: FocusEvent) {
 		const eTarget = e.relatedTarget as HTMLElement;
-		const focusedOutToValueStateMessage = eTarget?.querySelector(".ui5-valuestatemessage-root");
+		const focusedOutToValueStateMessage = eTarget?.shadowRoot!.querySelector(".ui5-valuestatemessage-root");
 
 		this.focused = false;
 
@@ -627,6 +632,10 @@ class TextArea extends UI5Element implements IFormElement {
 			return;
 		}
 
+		if (this.valueState === ValueState.None) {
+			return;
+		}
+
 		if (this.hasCustomValueState) {
 			return `${this.valueStateTypeMappings[this.valueState]}`.concat(" ", this.valueStateMessageText.map(el => el.textContent).join(" "));
 		}
@@ -635,7 +644,11 @@ class TextArea extends UI5Element implements IFormElement {
 	}
 
 	get valueStateDefaultText() {
-		return this.valueStateTextMappings[this.valueState];
+		if (this.valueState !== ValueState.None) {
+			return this.valueStateTextMappings[this.valueState];
+		}
+
+		return "";
 	}
 
 	get ariaInvalid() {
@@ -686,7 +699,6 @@ class TextArea extends UI5Element implements IFormElement {
 			"Information": TextArea.i18nBundle.getText(VALUE_STATE_INFORMATION),
 			"Error": TextArea.i18nBundle.getText(VALUE_STATE_ERROR),
 			"Warning": TextArea.i18nBundle.getText(VALUE_STATE_WARNING),
-			"None": "",
 		};
 	}
 
@@ -696,7 +708,6 @@ class TextArea extends UI5Element implements IFormElement {
 			"Information": TextArea.i18nBundle.getText(VALUE_STATE_TYPE_INFORMATION),
 			"Error": TextArea.i18nBundle.getText(VALUE_STATE_TYPE_ERROR),
 			"Warning": TextArea.i18nBundle.getText(VALUE_STATE_TYPE_WARNING),
-			"None": "",
 		};
 	}